+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Fix `TileJSON` init parameters
 * Rename `CustomFieldsTileJSON` to `ExtendedTileJSON`
 * Rename `ExtendedTileJSON.customFields` to `extendedFields`.
-* Refactor `EffectiveTileJSONFields` to inherit from `TileJSONFields`.
+* Combine `EffectiveTileJSONFields` and `TileJSONFields`.
 
 ## 0.1.1
 * Add more documentation